Move more functionality to util.py
Move class EpochField serialiser and class ListArgsSerializer from
parking/serialisers.py to util.py for future reuse.
